refactor: move collections components to a dedicated collections directory
