+++ src/components/Sidenav.svelte
@@ -4,7 +4,7 @@
     import IconCollection from 'lucide-svelte/icons/library';
     import Search from 'lucide-svelte/icons/search';
     import IconSearch from './icons/IconSearch.svelte';
-    import CollectionList from './Sidebar/CollectionList.svelte';
+    import CollectionList from './collections/CollectionList.svelte';
 
     let selectedTile = $state('icons'); // Default to 'icons'
