+++ src/components/collections/CollectionIcons.svelte
